--- src/app/app.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'my-app',
-    templateUrl: 'src/app/templates/app.component.html',
-    styleUrls: ['src/app/templates/app.component.css']
-})
-
-export class AppComponent { }
