+++ src/app/components/app/app.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'my-app',
+    templateUrl: 'src/app/templates/app/app.component.html',
+    styleUrls: ['src/app/templates/app/app.component.css']
+})
+
+export class AppComponent { }
