+++ src/app/components/footer/footer.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: '[app-footer]',
+    templateUrl: 'src/app/templates/footer/footer.component.html',
+    styleUrls: ['src/app/templates/footer/footer.component.css']
+})
+
+export class AppFooter { }
