+++ src/app/components/header/header.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: '[app-header]',
+    templateUrl: 'src/app/templates/header/header.component.html',
+    styleUrls: ['src/app/templates/header/header.component.css']
+})
+
+export class AppHeader { }
