--- src/app/footer.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: '[app-footer]',
-    templateUrl: 'src/app/templates/footer.component.html',
-    styleUrls: ['src/app/templates/footer.component.css']
-})
-
-export class AppFooter { }
