--- src/app/header.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: '[app-header]',
-    templateUrl: 'src/app/templates/header.component.html',
-    styleUrls: ['src/app/templates/header.component.css']
-})
-
-export class AppHeader { }
