+++ src/app/templates/app/app.component.css
