+++ src/app/templates/weather/weather.component.html
